pyproject.toml
[project]
name = "osnova"
version = "0.1.0"
description = "Decentralized truth network with Dunbar-capped trust rings and PARDES integrity"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pynacl>=1.5.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: end-to-end tests that spawn real uvicorn processes (deselect with -m 'not integration')",
]
